Added lots of tools
I added a few new tools. Biggest among them being a tool that creates assets for us so that all we need to do is add the item icons. The item icons aren't associated with the items themselves other than through the scene, not the editor in any way, otherwise they would be automatic aswell.

Another big thing is that I've  added is the character creator itself. It's not functional yet, but it's very close. I think a couple more days of work and it'll be operational.

Along the way I've created a CopyComponentUtil which copies component parameters and pastes them onto a component which is not of the same type.

I've also added a new CustomToggle object which has in my opinion a much better visualization than just a sprite turning on or off.

A panel manager was added for menus and such.

And a ForceImmediateLayoutUpdate tool was added to make sure the LayoutGroup component would layout properly on start. There is an option to make this component periodically update, which I recommend as in my experience swapping between panels can cause it to need to update again.

I'm too tired to explain the rest of the changes, many are self explanatory like the CCItemButton or CharacterItem.

These additions have all been in order to move the old character creation demo over to a proper system which we can use to start generating character assets.