Expanded character creator to allow user to modify a character
Character creation is about 75% of the way there. I need to get the face mutation to work again, and get the save and loading to work. Finally I'll be setting up a tag system so that each character can be designated as a particular character type. For example marking a policeman as "police".
I started work on getting the UICharacterCustomizerController cleaned up, and getting the components of it working.
Finally I setup the icons for the items. I started work on an automatic icon finder, but did not finish it and don't intend to for some time.