Character creation work
I setup the automatic sprite location to work properly on the CharacterItemList. It's still a little buggy, but it works allot better now.

The face customization was added, I didn't fully test it, but I think it's working. I'll be having some one else run a full test on it down the line when they create characters using the system.

Finally got a save setup ready for work with the characters, and will soon be able to properly save and load characters from the system. Once this is implemented I will be calling it a day, and moving on to more important features.